cli: Fix argument parsing on GNU
The command line tool assumes getopt(3) parses the arguments in order.
However, the GNU libc first permutes argv[] so that non-options (such as
"detokenize") are at the end.

The GNU getopt(3) can be forced into compatible mode either via
POSIXLY_CORRECT environment variable, or by starting the optstring with
a '+'. Let's do the latter -- it seems to not cause trouble on SunOS
either.

Tested on the following:
1.) Fedora Linux 40, x86_64
2.) SunOS sunos 5.11 tribblix-m34 i86pc i386 i86pc